homework

+++ Module305JDBC/Module305JDBC/src/main/java/org/example/ProductMain.java
@@ -48,7 +48,7 @@ public class ProductMain {
     }
 
     public void updateQuantityInStock(Product product, int quantity) {
-        product.setQuantityInStock((int) quantity);
+        product.setQuantityInStock((short) quantity);
         productDAO.update(product);
     }
